Avance
finalizando la integracion con la lista BD, falta con el arbol B
