+++ code/BDmodelo.py
@@ -1,0 +1,4 @@
+from listaDoble import *
+
+
+bd = ListaDOBLE()
