+++ code/llamadaMetodos.py
@@ -18,3 +18,22 @@ drop = dropDatabase('db4')
 print(drop)
 imp = showDatabases()
 print(imp)
+
+
+#--tablas
+print("\n-----")
+print(createTable("db1","table1",4))
+print(createTable("db1","table2",5))
+print(createTable("db2","table3",9))
+print(createTable("db2","table4",2))
+print(createTable("db3","table5",3))
+
+print(showTables("db1"))
+print(showTables("db2"))
+#print(extractTable("db1", "table1"))
+#print(extractRangeTable("bd1","tabla1",2,0,2))
+print(alterTable("db1","table1","tableNew"))
+#alterAddColumn(database, table, columnNumber) :
+#print(alterDropColumn("db1", "table1", 2))
+print(dropTable("db1","table1"))
+print(dropTable("db1","table2"))
